requirements.txt
flask==1.1.2
beautifulsoup4==4.9.3
pyyaml==5.3.1
marko==0.10.0
